requirements.txt
pulumi
pulumi-aws
flask
python-dotenv

